+++ custard/utils.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 
 import importlib
+import functools
+import warnings
 
 
 #==============================================================================
@@ -10,3 +12,14 @@ def import_class(name):
     for comp in components[1:]:
         mod = getattr(mod, comp)
     return mod
+
+
+#==============================================================================
+def deprecate(msg, klass=PendingDeprecationWarning):
+    def decorator(func):
+        @functools.wraps(func)
+        def wrapper(*args, **kwargs):
+            warnings.warn(msg, klass, stacklevel=2)
+            return func(*args, **kwargs)
+        return wrapper
+    return decorator